+++ core/src/window.rs
@@ -280,7 +280,7 @@ impl Window {
     if self.painter.borrow().viewport().size != size {
       let tree = self.tree_mut();
       let root = tree.root();
-      tree.dirty_marker().mark(root);
+      tree.dirty_marker().mark(root, DirtyPhase::Layout);
       tree.store.remove(root);
       let mut painter = self.painter.borrow_mut();
       painter.set_viewport(Rect::from_size(size));